requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.12.3
cryptography==2.6.1
Fabric3==1.14.post1
invoke==1.2.0
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
six==1.12.0
virtualenv==16.4.3
